New branch for updates and external contribs.
While the master branch of this repository will be kept freezed so that
it is consistent with the original paper, this new branch will be
updated with further observations and contributions. For example, we
could update it to use recent features from new compilers or we
could incorporate suggestions from external users.